Add dataset argument
