+++ vqvae.py
@@ -1,149 +1,199 @@
 import numpy as np
-import time
-from pathlib import Path
-
 import torch
 import torch.nn.functional as F
-from torchvision import datasets, transforms
-from torchvision.utils import save_image
-from torch.distributions.normal import Normal
+from torchvision import transforms, datasets
+from torchvision.utils import save_image, make_grid
 
 from modules import VectorQuantizedVAE, to_scalar
+from datasets import MiniImagenet
 
+from tensorboardX import SummaryWriter
 
-BATCH_SIZE = 32
-N_EPOCHS = 100
-PRINT_INTERVAL = 100
-DATASET = 'CIFAR10'  # CIFAR10 | MNIST | FashionMNIST
-NUM_WORKERS = 4
+def train(data_loader, model, optimizer, args, writer):
+    for images, _ in data_loader:
+        images = images.to(args.device)
 
-INPUT_DIM = 3  # 3 (RGB) | 1 (Grayscale)
-DIM = 256
-K = 512
-LAMDA = 1
-LR = 1e-3
+        optimizer.zero_grad()
+        x_tilde, z_e_x, z_q_x = model(images)
 
-DEVICE = torch.device('cuda')  # torch.device('cpu')
-
-
-# Create directories if they don't exist
-Path('models').mkdir(exist_ok=True)
-Path('samples').mkdir(exist_ok=True)
-
-preproc_transform = transforms.Compose([
-    transforms.ToTensor(),
-    transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
-])
-train_loader = torch.utils.data.DataLoader(
-    eval('datasets.'+DATASET)(
-        '../data/{}/'.format(DATASET), train=True, download=True,
-        transform=preproc_transform,
-    ), batch_size=BATCH_SIZE, shuffle=False,
-    num_workers=NUM_WORKERS, pin_memory=True
-)
-test_loader = torch.utils.data.DataLoader(
-    eval('datasets.'+DATASET)(
-        '../data/{}/'.format(DATASET), train=False,
-        transform=preproc_transform,
-    ), batch_size=BATCH_SIZE, shuffle=False,
-    num_workers=NUM_WORKERS, pin_memory=True
-)
-
-model = VectorQuantizedVAE(INPUT_DIM, DIM, K).to(DEVICE)
-print(model)
-opt = torch.optim.Adam(model.parameters(), lr=LR, amsgrad=True)
-
-
-def train():
-    train_loss = []
-    for batch_idx, (x, _) in enumerate(train_loader):
-        start_time = time.time()
-        x = x.to(DEVICE)
-
-        opt.zero_grad()
-
-        x_tilde, z_e_x, z_q_x = model(x)
-        z_q_x.retain_grad()
-
-        loss_recons = F.mse_loss(x_tilde, x)
-        loss_recons.backward(retain_graph=True)
-
-        # Straight-through estimator
-        z_e_x.backward(z_q_x.grad, retain_graph=True)
-
+        # Reconstruction loss
+        loss_recons = F.mse_loss(x_tilde, images)
         # Vector quantization objective
-        model.codebook.zero_grad()
         loss_vq = F.mse_loss(z_q_x, z_e_x.detach())
-        loss_vq.backward(retain_graph=True)
-
         # Commitment objective
-        loss_commit = LAMDA * F.mse_loss(z_e_x, z_q_x.detach())
-        loss_commit.backward()
-        opt.step()
+        loss_commit = F.mse_loss(z_e_x, z_q_x.detach())
 
-        N = x.numel()
-        nll = Normal(x_tilde, torch.ones_like(x_tilde)).log_prob(x)
-        log_px = nll.sum() / N + np.log(128) - np.log(K * 2)
-        log_px /= np.log(2)
+        loss = loss_recons + loss_vq + args.beta * loss_commit
+        loss.backward()
 
-        train_loss.append(
-            [log_px.item()] + to_scalar([loss_recons, loss_vq])
-        )
+        # Logs
+        writer.add_scalar('loss/train/reconstruction', loss_recons.item(), args.steps)
+        writer.add_scalar('loss/train/quantization', loss_vq.item(), args.steps)
 
-        if (batch_idx + 1) % PRINT_INTERVAL == 0:
-            print('\tIter [{}/{} ({:.0f}%)]\tLoss: {} Time: {}'.format(
-                batch_idx * len(x), len(train_loader.dataset),
-                PRINT_INTERVAL * batch_idx / len(train_loader),
-                np.asarray(train_loss)[-PRINT_INTERVAL:].mean(0),
-                time.time() - start_time
-            ))
+        optimizer.step()
+        args.steps += 1
 
+def test(data_loader, model, args, writer):
+    with torch.no_grad():
+        loss_recons, loss_vq = 0., 0.
+        for images, _ in data_loader:
+            images = images.to(args.device)
+            x_tilde, z_e_x, z_q_x = model(images)
+            loss_recons += F.mse_loss(x_tilde, images)
+            loss_vq += F.mse_loss(z_q_x, z_e_x)
 
-def test():
-    start_time = time.time()
-    val_loss = []
-    for batch_idx, (x, _) in enumerate(test_loader):
-        x = x.to(DEVICE)
-        x_tilde, z_e_x, z_q_x = model(x)
-        loss_recons = F.mse_loss(x_tilde, x)
-        loss_vq = F.mse_loss(z_q_x, z_e_x.detach())
-        val_loss.append(to_scalar([loss_recons, loss_vq]))
+        loss_recons /= len(data_loader)
+        loss_vq /= len(data_loader)
 
-    print('\nValidation Completed!\tLoss: {} Time: {:5.3f}'.format(
-        np.asarray(val_loss).mean(0),
-        time.time() - start_time
-    ))
-    return np.asarray(val_loss).mean(0)
+    # Logs
+    writer.add_scalar('loss/test/reconstruction', loss_recons.item(), args.steps)
+    writer.add_scalar('loss/test/quantization', loss_vq.item(), args.steps)
 
+    return loss_recons.item(), loss_vq.item()
 
-def generate_samples():
-    x, _ = test_loader.__iter__().next()
-    x = x[:32].to(DEVICE)
-    x_tilde, _, _ = model(x)
+def generate_samples(images, model, args):
+    with torch.no_grad():
+        images = images.to(args.device)
+        x_tilde, _, _ = model(images)
+    return x_tilde
 
-    x_cat = torch.cat([x, x_tilde], 0)
-    images = (x_cat.cpu().data + 1) / 2
+def main(args):
+    writer = SummaryWriter('./logs/{0}'.format(args.output_folder))
+    save_filename = './models/{0}'.format(args.output_folder)
 
-    save_image(
-        images,
-        'samples/vqvae_reconstructions_{}.png'.format(DATASET),
-        nrow=8
-    )
+    if args.dataset in ['mnist', 'fashion-mnist', 'cifar10']:
+        transform = transforms.Compose([
+            transforms.ToTensor(),
+            transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
+        ])
+        if args.dataset == 'mnist':
+            # Define the train & test datasets
+            train_dataset = datasets.MNIST(args.data_folder, train=True,
+                download=True, transform=transform)
+            test_dataset = datasets.MNIST(args.data_folder, train=False,
+                transform=transform)
+            num_channels = 1
+        elif args.dataset == 'fashion-mnist':
+            # Define the train & test datasets
+            train_dataset = datasets.FashionMNIST(args.data_folder,
+                train=True, download=True, transform=transform)
+            test_dataset = datasets.FashionMNIST(args.data_folder,
+                train=False, transform=transform)
+            num_channels = 1
+        elif args.dataset == 'cifar10':
+            # Define the train & test datasets
+            train_dataset = datasets.CIFAR10(args.data_folder,
+                train=True, download=True, transform=transform)
+            test_dataset = datasets.CIFAR10(args.data_folder,
+                train=False, transform=transform)
+            num_channels = 3
+        valid_dataset = test_dataset
+    elif args.dataset == 'miniimagenet':
+        transform = transforms.Compose([
+            transforms.RandomResizedCrop(128),
+            transforms.ToTensor(),
+            transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
+        ])
+        # Define the train, valid & test datasets
+        train_dataset = MiniImagenet(args.data_folder, train=True,
+            download=True, transform=transform)
+        valid_dataset = MiniImagenet(args.data_folder, valid=True,
+            download=True, transform=transform)
+        test_dataset = MiniImagenet(args.data_folder, test=True,
+            download=True, transform=transform)
+        num_channels = 3
 
+    # Define the data loaders
+    train_loader = torch.utils.data.DataLoader(train_dataset,
+        batch_size=args.batch_size, shuffle=False,
+        num_workers=args.num_workers, pin_memory=True)
+    valid_loader = torch.utils.data.DataLoader(valid_dataset,
+        batch_size=args.batch_size, shuffle=False, drop_last=True,
+        num_workers=args.num_workers, pin_memory=True)
+    test_loader = torch.utils.data.DataLoader(test_dataset,
+        batch_size=16, shuffle=True)
 
-BEST_LOSS = 999
-LAST_SAVED = -1
-for epoch in range(1, N_EPOCHS):
-    print("Epoch {}:".format(epoch))
-    train()
-    cur_loss, _ = test()
+    # Fixed images for Tensorboard
+    fixed_images, _ = next(iter(test_loader))
+    fixed_grid = make_grid(fixed_images, nrow=8, range=(-1, 1), normalize=True)
+    writer.add_image('original', fixed_grid, 0)
 
-    if cur_loss <= BEST_LOSS:
-        BEST_LOSS = cur_loss
-        LAST_SAVED = epoch
-        print("Saving model!")
-        torch.save(model.state_dict(), 'models/{}_vqvae.pt'.format(DATASET))
-    else:
-        print("Not saving model! Last saved: {}".format(LAST_SAVED))
+    model = VectorQuantizedVAE(num_channels, args.hidden_size, args.k).to(args.device)
+    optimizer = torch.optim.Adam(model.parameters(), lr=args.lr)
 
-    generate_samples()
+    # Generate the samples first once
+    reconstruction = generate_samples(fixed_images, model, args)
+    grid = make_grid(reconstruction.cpu(), nrow=8, range=(-1, 1), normalize=True)
+    writer.add_image('reconstruction', grid, 0)
+
+    best_loss = -1.
+    for epoch in range(args.num_epochs):
+        train(train_loader, model, optimizer, args, writer)
+        loss, _ = test(valid_loader, model, args, writer)
+
+        reconstruction = generate_samples(fixed_images, model, args)
+        grid = make_grid(reconstruction.cpu(), nrow=8, range=(-1, 1), normalize=True)
+        writer.add_image('reconstruction', grid, epoch + 1)
+
+        if (epoch == 0) or (loss < best_loss):
+            best_loss = loss
+            with open('{0}/best.pt'.format(save_filename), 'wb') as f:
+                torch.save(model.state_dict(), f)
+        with open('{0}/model_{1}.pt'.format(save_filename, epoch + 1), 'wb') as f:
+            torch.save(model.state_dict(), f)
+
+if __name__ == '__main__':
+    import argparse
+    import os
+    import multiprocessing as mp
+
+    parser = argparse.ArgumentParser(description='VQ-VAE')
+
+    # General
+    parser.add_argument('--data-folder', type=str,
+        help='name of the data folder')
+    parser.add_argument('--dataset', type=str,
+        help='name of the dataset (mnist, fashion-mnist, cifar10, miniimagenet)')
+
+    # Latent space
+    parser.add_argument('--hidden-size', type=int, default=256,
+        help='size of the latent vectors (default: 256)')
+    parser.add_argument('--k', type=int, default=512,
+        help='number of latent vectors (default: 512)')
+
+    # Optimization
+    parser.add_argument('--batch-size', type=int, default=128,
+        help='batch size (default: 128)')
+    parser.add_argument('--num-epochs', type=int, default=100,
+        help='number of epochs (default: 100)')
+    parser.add_argument('--lr', type=float, default=2e-4,
+        help='learning rate for Adam optimizer (default: 2e-4)')
+    parser.add_argument('--beta', type=float, default=1.0,
+        help='contribution of commitment loss, between 0.1 and 2.0 (default: 1.0)')
+
+    # Miscellaneous
+    parser.add_argument('--output-folder', type=str, default='vqvae',
+        help='name of the output folder (default: vqvae)')
+    parser.add_argument('--num-workers', type=int, default=mp.cpu_count() - 1,
+        help='number of workers for trajectories sampling (default: {0})'.format(mp.cpu_count() - 1))
+    parser.add_argument('--device', type=str, default='cpu',
+        help='set the device (cpu or cuda, default: cpu)')
+
+    args = parser.parse_args()
+
+    # Create logs and models folder if they don't exist
+    if not os.path.exists('./logs'):
+        os.makedirs('./logs')
+    if not os.path.exists('./models'):
+        os.makedirs('./models')
+    # Device
+    args.device = torch.device(args.device
+        if torch.cuda.is_available() else 'cpu')
+    # Slurm
+    if 'SLURM_JOB_ID' in os.environ:
+        args.output_folder += '-{0}'.format(os.environ['SLURM_JOB_ID'])
+    if not os.path.exists('./models/{0}'.format(args.output_folder)):
+        os.makedirs('./models/{0}'.format(args.output_folder))
+    args.steps = 0
+
+    main(args)
